refactor: DTOs
