+++ src/main/java/microservice/pacientes/controller/PacienteControllerImpl.java
@@ -1,12 +1,8 @@
 package microservice.pacientes.controller;
 
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import microservice.pacientes.dto.PacienteRequestDTO;
-import microservice.pacientes.dto.PacienteResponseDTO;
-import microservice.pacientes.dto.PacienteUpdateDTO;
+import microservice.pacientes.controller.dto.PacienteRequestDTO;
+import microservice.pacientes.controller.dto.PacienteResponseDTO;
+import microservice.pacientes.controller.dto.PacienteUpdateDTO;
 import microservice.pacientes.model.Paciente;
 import microservice.pacientes.service.PacienteService;
 import microservice.pacientes.util.PacienteResponseMapper;
