+++ src/main/java/microservice/pacientes/controller/dto/PacienteResponseDTO.java
@@ -1,18 +1,17 @@
-package microservice.pacientes.dto;
+package microservice.pacientes.controller.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@AllArgsConstructor
-@NoArgsConstructor
 @Getter
 @Setter
-public class PacienteUpdateDTO {
+@AllArgsConstructor
+@NoArgsConstructor
+public class PacienteResponseDTO {
+    private String dni;
     private String nombre;
     private String apellido;
     private String obraSocial;
-    private String email;
-    private String telefono;
 }
