+++ src/main/java/microservice/pacientes/service/PacienteService.java
@@ -1,8 +1,7 @@
 package microservice.pacientes.service;
 
-import microservice.pacientes.dto.PacienteRequestDTO;
-import microservice.pacientes.dto.PacienteResponseDTO;
-import microservice.pacientes.dto.PacienteUpdateDTO;
+import microservice.pacientes.controller.dto.PacienteRequestDTO;
+import microservice.pacientes.controller.dto.PacienteUpdateDTO;
 import microservice.pacientes.exception.PacienteDuplicadoException;
 import microservice.pacientes.exception.PacienteNoEncontradoException;
 import microservice.pacientes.model.Paciente;
